moved each component to separate
