+++ src/components/styledComponents/StyledTableCell.tsx
@@ -1,0 +1,19 @@
+import { TableCell, tableCellClasses } from '@mui/material';
+import { styled } from '@mui/system';
+
+export const StyledTableCell = styled(TableCell)(({theme}) => ({
+    fontSize: 16,
+    textAlign: 'center',
+    color: theme.palette.primary.light,
+    fontFamily: "'Comic Sans MS', sans-serif",
+    padding: 10,
+    borderBottom: '1px solid #ddd',
+    
+    [`&.${tableCellClasses.head}`]: {
+      backgroundColor: theme.palette.secondary.dark,
+      color: theme.palette.secondary.light,
+    },
+    [`&.${tableCellClasses.body}`]: {
+      fontSize: 14,
+    },
+}));
