+++ src/components/styledComponents/StyledTableCellMobile.tsx
@@ -1,0 +1,6 @@
+import { StyledTableCell } from './StyledTableCell';
+import { styled } from '@mui/system';
+
+export const StyledTableCellMobile = styled(StyledTableCell)(() => ({
+  textAlign: 'left',
+}));
